+++ transport/hub/kcpv/config_json.go
@@ -1,0 +1,1 @@
+package kcpv
